Release v1.1.0: Implement ASIL-D safe state escalation, CodeQL, and gPTP fixes
Release v1.1.0 of the norxs Deterministic TSN Zonal Backbone.

This release fulfills critical ASIL-D safety state mechanisms, hardens the gPTP synchronization engine, and deepens our automated security and compliance pipelines.

Key additions and improvements:
* **ASIL-D Safe State Escalation:** Fully implemented the `kDegraded → kSafeState` transition. The orchestrator now actively revokes Network Calculus guarantees and applies a TC7-exclusive fail-safe schedule across all ports if a hardware GCL integrity fault persists.
* **gPTP Convergence Hardening (OBS-001):** Introduced a PHC-independent monotonic clock to reliably drive the gPTP convergence watchdog. This ensures mathematically proven `kFault` escalation for perpetually isolated slave nodes.
* **Hardware GCL Alignment (OBS-002):** Widened `ScheduleParams::entryCount` to `u16`, enabling full access to the 256-entry GCL SRAM depth of NXP SJA1110 and Marvell 88Q5050 ASICs.
* **Verification Depth:** Achieved 96.1% overall structural line coverage (113/113 tests passing). Added 5 critical holdover runtime-path tests validating PTP sync-silence recovery and TCXO drift compensation.
* **Security Tooling & Compliance:** Integrated GitHub CodeQL semantic security analysis for continuous ISO/IEC 18974 §3.1.5 compliance. Added OSPO/Safety dual-review via `CODEOWNERS` and regenerated the SPDX 2.3 SBOM.
* **End-to-End Demo:** Shipped `zonal_gateway_demo.cpp`, a runnable lifecycle demo simulating fault injection and audit-trail generation against a switch ASIC.

+++ include/norxs/tsn/TsnTypes.hpp
@@ -303,7 +303,7 @@ struct ScheduleParams
     u64              cycleTimeNs;                         ///< Repeating period [ns]
     PtpTimestamp     baseTime;                            ///< Absolute PTP cycle start
     u32              maxSafetyFrameSizeBytes;             ///< Largest expected TC7 frame
-    u8               entryCount;                          ///< Valid GCL entries [1, 256]
+    u16              entryCount;                          ///< Valid GCL entries [1, 256]
     std::array<GclEntry, kMaxGclEntries> gcl;            ///< Gate Control List
 };
